requirements.txt
SQLAlchemy==2.0.46
alembic==1.18.3
fastapi==0.128.1
pwdlib==0.3.0
PyJWT==2.11.0
python-decouple==3.8
